+++ app/smartbrain_3000/web/_app/immutable/chunks/aUuKMASR.js
@@ -1,1 +1,0 @@
-import{q as e}from"./Dcti9EvM.js";import{n as t}from"./BgU3LVSf.js";var n=e({count:0});async function r(){try{n.count=(await t.unseenScheduleUpdates()).count}catch{}}export{n,r as t};